fixed combine reducer
