+++ src/redux/store.js
@@ -6,31 +6,31 @@ import { createStore, combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
 
 import uiReducer from '../ui/redux/uiReducer';
-import quizData from '../ui/redux/reducers/quizData';
+import quizDataReducer from '../ui/redux/reducers/quizDataReducer';
 
-// const rootReducer = combineReducers({
-//   form: formReducer,
-//   ui: uiReducer,
-//   quiz: quizData,
-// });
+const rootReducer = combineReducers({
+  form: formReducer,
+  ui: uiReducer,
+  quizData: quizDataReducer,
+});
 
-const initialState = {
-  quizData: {}
-}
+// const initialState = {
+//   quizData: {}
+// }
 
-const reducer = (state = initialState, action) => {
-  console.log('reducer running', action);
+// const reducer = (state = initialState, action) => {
+//   console.log('reducer running', action);
 
-  switch (action.type) {
-    case 'SAVE_QUIZ_DATA':
-      return Object.assign({}, state, { quizData: action.payload });
-    default:
-      return state;
-  }
-}
+//   switch (action.type) {
+//     case 'SAVE_QUIZ_DATA':
+//       return Object.assign({}, state, { quizData: action.payload });
+//     default:
+//       return state;
+//   }
+// }
 
-// const store = createStore(rootReducer);
+const store = createStore(rootReducer);
 
-const store = createStore(reducer);
+// const store = createStore(reducer);
 
 export default store;
